Extract shared slash command mixer
Share slash command mixer construction so GUI and TUI can build separate SearchMixer instances from the same source composition. Add a TUI-aware slash data source constructor and export the shared pieces through tui_export.
