+++ app/src/terminal/input/slash_commands/cloud_mode_v2_view.rs
@@ -24,10 +24,10 @@ use crate::search::slash_command_menu::static_commands::commands::COMMAND_REGIST
 use crate::terminal::input::buffer_model::{InputBufferModel, InputBufferUpdateEvent};
 use crate::terminal::input::inline_menu::{styles as inline_styles, QueryResultRendererExt as _};
 use crate::terminal::input::slash_command_model::{SlashCommandEntryState, SlashCommandModel};
-use crate::terminal::input::slash_commands::view::{slash_command_query, CloseReason};
+use crate::terminal::input::slash_commands::view::CloseReason;
 use crate::terminal::input::slash_commands::{
-    saved_prompts_data_source, AcceptSlashCommandOrSavedPrompt, SlashCommandDataSource,
-    SlashCommandsEvent, UpdatedActiveCommands, ZeroStateDataSource,
+    saved_prompts_data_source, slash_command_query, AcceptSlashCommandOrSavedPrompt,
+    GuiSlashCommandDataSource, GuiZeroStateDataSource, SlashCommandsEvent, UpdatedActiveCommands,
 };
 use crate::terminal::input::suggestions_mode_model::{
     InputSuggestionsModeEvent, InputSuggestionsModeModel,
@@ -201,7 +201,7 @@ pub struct CloudModeV2SlashCommandView {
 impl CloudModeV2SlashCommandView {
     pub fn new(
         slash_command_model: &ModelHandle<SlashCommandModel>,
-        slash_commands_source: ModelHandle<SlashCommandDataSource>,
+        slash_commands_source: ModelHandle<GuiSlashCommandDataSource>,
         suggestions_mode_model: ModelHandle<InputSuggestionsModeModel>,
         input_buffer_model: ModelHandle<InputBufferModel>,
         ctx: &mut ViewContext<Self>,
@@ -218,7 +218,7 @@ impl CloudModeV2SlashCommandView {
         );
 
         let zero_state_source =
-            ctx.add_model(|_| ZeroStateDataSource::new(&slash_commands_source, true));
+            ctx.add_model(|_| GuiZeroStateDataSource::new(&slash_commands_source));
         let saved_prompts_source = saved_prompts_data_source();
 
         let mixer = ctx.add_model(|ctx| {
