+++ app/src/terminal/input/slash_commands/data_source/saved_prompts.rs
@@ -6,7 +6,6 @@ use warp_core::ui::appearance::Appearance;
 use warpui::fonts::FamilyId;
 use warpui::{AppContext, SingletonEntity};
 
-use super::{AcceptSlashCommandOrSavedPrompt, InlineItem};
 use crate::cloud_object::model::persistence::CloudModel;
 use crate::cloud_object::CloudObject;
 use crate::search::async_snapshot_data_source::AsyncSnapshotDataSource;
@@ -15,6 +14,7 @@ use crate::search::mixer::{BoxFuture, DataSourceRunErrorWrapper};
 use crate::search::FuzzyMatchWorkflowResult;
 use crate::server::ids::SyncId;
 use crate::settings::AISettings;
+use crate::terminal::input::slash_commands::{AcceptSlashCommandOrSavedPrompt, InlineItem};
 use crate::workflows::CloudWorkflowModel;
 
 pub(super) struct SavedPromptCandidate {
