+++ app/src/terminal/input/slash_commands/data_source/zero_state.rs
@@ -6,35 +6,30 @@ use crate::ai::skills::SkillManager;
 use crate::cloud_object::model::persistence::CloudModel;
 use crate::search::data_source::{Query, QueryResult};
 use crate::search::mixer::DataSourceRunErrorWrapper;
-use crate::search::slash_command_menu::static_commands::commands;
 use crate::search::SyncDataSource;
 use crate::settings::AISettings;
 use crate::terminal::input::slash_commands::{
-    AcceptSlashCommandOrSavedPrompt, InlineItem, SlashCommandDataSource,
+    AcceptSlashCommandOrSavedPrompt, GuiSlashCommandDataSource, InlineItem,
+    TuiSlashCommandDataSource,
 };
 
-pub struct ZeroStateDataSource {
-    slash_command_data_source: ModelHandle<SlashCommandDataSource>,
-    is_cloud_mode_v2: bool,
+pub struct GuiZeroStateDataSource {
+    slash_command_data_source: ModelHandle<GuiSlashCommandDataSource>,
 }
 
-impl ZeroStateDataSource {
-    pub fn new(
-        slash_command_data_source: &ModelHandle<SlashCommandDataSource>,
-        is_cloud_mode_v2: bool,
-    ) -> Self {
+impl GuiZeroStateDataSource {
+    pub fn new(slash_command_data_source: &ModelHandle<GuiSlashCommandDataSource>) -> Self {
         Self {
             slash_command_data_source: slash_command_data_source.clone(),
-            is_cloud_mode_v2,
         }
     }
 }
 
-impl Entity for ZeroStateDataSource {
+impl Entity for GuiZeroStateDataSource {
     type Event = ();
 }
 
-impl SyncDataSource for ZeroStateDataSource {
+impl SyncDataSource for GuiZeroStateDataSource {
     type Action = AcceptSlashCommandOrSavedPrompt;
 
     fn run_query(
@@ -46,68 +41,17 @@ impl SyncDataSource for ZeroStateDataSource {
             return Ok(vec![]);
         }
 
-        // This is kind of a convoluted way to explicitly order these commands after all others.
-        //
-        // DataSource implementations must return highest priority items last (results sorted in
-        // ascending order of priority).
-        //
-        // The results construction below basically orders all active commands, sorted
-        // alphabetically, except for the commands in this vec, which are explicitly appended
-        // to all the other alphabetically sorted commands, in this order.
-        let prioritized_commands = vec![
-            &*commands::CREATE_ENVIRONMENT,
-            &*commands::EDIT,
-            &commands::CONVERSATIONS,
-            &commands::PROMPTS,
-            &*commands::PLAN,
-            &commands::AGENT,
-        ];
+        let source = self.slash_command_data_source.as_ref(app);
+        let is_cloud_mode_v2 = source.is_cloud_mode_v2();
+        let mut results = source.core().ordered_zero_state_commands(app);
 
-        let mut active_prioritized_commands = vec![];
-        let mut results = vec![];
-
-        for (active_command_id, active_command) in self
-            .slash_command_data_source
-            .as_ref(app)
-            .active_commands()
-            .sorted_by_key(|(_, command)| std::cmp::Reverse(&command.name))
-        {
-            if prioritized_commands
-                .iter()
-                .any(|prioritized_command| prioritized_command.name == active_command.name)
-            {
-                active_prioritized_commands.push((active_command_id, active_command));
-            } else {
-                results.push(
-                    InlineItem::from_slash_command(active_command_id, active_command, app)
-                        .with_compact_layout(self.is_cloud_mode_v2)
-                        .into(),
-                );
-            }
-        }
-
-        for prioritized_command in prioritized_commands {
-            if let Some((id, command)) = active_prioritized_commands
-                .iter()
-                .find(|(_, active_command)| active_command.name == prioritized_command.name)
-            {
-                results.push(
-                    InlineItem::from_slash_command(id, command, app)
-                        .with_compact_layout(self.is_cloud_mode_v2)
-                        .into(),
-                );
-            }
-        }
-
-        if self.is_cloud_mode_v2
+        if is_cloud_mode_v2
             && FeatureFlag::ListSkills.is_enabled()
             && AISettings::as_ref(app).is_any_ai_enabled(app)
         {
-            let slash_command_data_source = self.slash_command_data_source.as_ref(app);
-            let cli_agent_providers = slash_command_data_source.active_cli_agent_providers(app);
-            let active_session = slash_command_data_source
-                .active_session_for_v2_zero_state()
-                .as_ref(app);
+            let core = source.core();
+            let cli_agent_providers = core.active_cli_agent_providers(app);
+            let active_session = core.active_session().as_ref(app);
             let cwd = active_session.current_working_directory_location(app);
             let skill_manager_handle = SkillManager::handle(app);
             let skill_manager = skill_manager_handle.as_ref(app);
@@ -123,15 +67,11 @@ impl SyncDataSource for ZeroStateDataSource {
                     }
                     skill.provider = skill_manager.best_supported_provider(&skill, providers);
                 }
-                results.push(
-                    InlineItem::from_skill(&skill, app)
-                        .with_compact_layout(self.is_cloud_mode_v2)
-                        .into(),
-                );
+                results.push(InlineItem::from_skill(&skill, app));
             }
         }
 
-        if self.is_cloud_mode_v2 && AISettings::as_ref(app).is_any_ai_enabled(app) {
+        if is_cloud_mode_v2 && AISettings::as_ref(app).is_any_ai_enabled(app) {
             let saved_prompts: Vec<_> = CloudModel::as_ref(app)
                 .get_all_active_workflows()
                 .filter(|cw| cw.model().data.is_agent_mode_workflow())
@@ -144,14 +84,52 @@ impl SyncDataSource for ZeroStateDataSource {
                 })
                 .collect();
             for saved_prompt in saved_prompts {
-                results.push(
-                    InlineItem::from_saved_prompt(saved_prompt, app)
-                        .with_compact_layout(self.is_cloud_mode_v2)
-                        .into(),
-                );
+                results.push(InlineItem::from_saved_prompt(saved_prompt, app));
             }
         }
 
-        Ok(results)
+        Ok(results
+            .into_iter()
+            .map(|item| item.with_compact_layout(is_cloud_mode_v2).into())
+            .collect())
+    }
+}
+
+pub struct TuiZeroStateDataSource {
+    slash_command_data_source: ModelHandle<TuiSlashCommandDataSource>,
+}
+
+impl TuiZeroStateDataSource {
+    pub fn new(slash_command_data_source: &ModelHandle<TuiSlashCommandDataSource>) -> Self {
+        Self {
+            slash_command_data_source: slash_command_data_source.clone(),
+        }
+    }
+}
+
+impl Entity for TuiZeroStateDataSource {
+    type Event = ();
+}
+
+impl SyncDataSource for TuiZeroStateDataSource {
+    type Action = AcceptSlashCommandOrSavedPrompt;
+
+    fn run_query(
+        &self,
+        query: &Query,
+        app: &warpui::AppContext,
+    ) -> Result<Vec<QueryResult<Self::Action>>, DataSourceRunErrorWrapper> {
+        if !query.text.is_empty() {
+            return Ok(vec![]);
+        }
+
+        Ok(self
+            .slash_command_data_source
+            .as_ref(app)
+            .core()
+            .ordered_zero_state_commands(app)
+            .into_iter()
+            .map(Into::into)
+            .collect())
     }
 }
